refactor: rename scripts
