+++ LOGGER/singleton.py
